reorganise discount mutations
