+++ saleor/graphql/discount/mutations/sale_base_catalogue.py
@@ -1,0 +1,21 @@
+import graphene
+
+from ..types import Sale
+from .sale_base_discount_catalogue import BaseDiscountCatalogueMutation
+from .voucher_add_catalogues import CatalogueInput
+
+
+class SaleBaseCatalogueMutation(BaseDiscountCatalogueMutation):
+    sale = graphene.Field(
+        Sale, description="Sale of which catalogue IDs will be modified."
+    )
+
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of a sale.")
+        input = CatalogueInput(
+            required=True,
+            description="Fields required to modify catalogue IDs of sale.",
+        )
+
+    class Meta:
+        abstract = True
