+++ saleor/graphql/product/tests/test_product_discounted_price.py
@@ -178,7 +178,7 @@ def test_collection_remove_products_updates_discounted_price(
 
 @freeze_time("2010-05-31 12:00:01")
 @patch(
-    "saleor.graphql.discount.mutations"
+    "saleor.graphql.discount.mutations.sale_create"
     ".update_products_discounted_prices_of_discount_task"
 )
 def test_sale_create_updates_products_discounted_prices(
@@ -231,7 +231,7 @@ def test_sale_create_updates_products_discounted_prices(
 
 
 @patch(
-    "saleor.graphql.discount.mutations"
+    "saleor.graphql.discount.mutations.sale_create"
     ".update_products_discounted_prices_of_discount_task"
 )
 def test_sale_update_updates_products_discounted_prices(
@@ -268,7 +268,7 @@ def test_sale_update_updates_products_discounted_prices(
 
 
 @patch(
-    "saleor.graphql.discount.mutations"
+    "saleor.graphql.discount.mutations.sale_create"
     ".update_products_discounted_prices_of_discount_task"
 )
 def test_sale_delete_updates_products_discounted_prices(
@@ -305,7 +305,7 @@ def test_sale_delete_updates_products_discounted_prices(
 
 
 @patch(
-    "saleor.graphql.discount.mutations"
+    "saleor.graphql.discount.mutations.sale_base_discount_catalogue"
     ".update_products_discounted_prices_of_catalogues_task"
 )
 def test_sale_add_catalogues_updates_products_discounted_prices(
@@ -365,7 +365,7 @@ def test_sale_add_catalogues_updates_products_discounted_prices(
 
 
 @patch(
-    "saleor.graphql.discount.mutations"
+    "saleor.graphql.discount.mutations.sale_base_discount_catalogue"
     ".update_products_discounted_prices_of_catalogues_task"
 )
 def test_sale_remove_catalogues_updates_products_discounted_prices(
